+++ test/support/bufferObject.c
@@ -13,8 +13,7 @@ static stc_bffr_t0 stc_bffr_pv_buf0;
 stc_bffr_t0* const p_stc_bffr_g_buf0 = &stc_bffr_pv_buf0;
 
 /* Private storages for buffer objects */
-static cirquStrgElem_t0 a_cirquStrgElem_pv_buf0Strg[BUF0STRGSIZE] =
-    { (uint8_t)0U };
+static cirquStrgElem_t0 a_cirquStrgElem_pv_buf0Strg[BUF0STRGSIZE] = {0U};
 
 /* OPERATIONS
  * ==========
@@ -22,8 +21,7 @@ static cirquStrgElem_t0 a_cirquStrgElem_pv_buf0Strg[BUF0STRGSIZE] =
 
 void fn_bffrObj_buf0Ctor(void)
 {
-    fn_bffr_ini0(&stc_bffr_pv_buf0, a_cirquStrgElem_pv_buf0Strg,
-                (uint8_t)BUF0STRGSIZE);
+    fn_bffr_ini0(&stc_bffr_pv_buf0, a_cirquStrgElem_pv_buf0Strg, BUF0STRGSIZE);
 
     return;
 }
